musdk: workaround for hif slot allocation for working threads
The purpose of this workaround is to allocate and init hif slots
only for threads that actually use it.
As we saw from OFP application, each thread created by app by using
odph_linux_pthread_create() ODP API function calls to local_init of
platform driver even if it's not going to use pktio interfaces.
For example: control thread or cli thread.
In this case we don't have any indication if we need to allocate
hif for this thread.

Change-Id: I8b1114b9b5368695e5f9d8201184c6a80f0678fa
Signed-off-by: nsamsono <nsamsono@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39812
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40256