tests: add scripts for running musdk validation tests
The original ODP validation tests, which are located under test/common_plat/validation/api, can be run independently.
The scripts below are only for convenience to be able to run a batch of the ODP tests and also be able to add our own tests.
All the files were copied from linux-generic, but some of them were modified.
* pktio_run.sh currently runs only original PKTIO test. Once we will add new tests this is the place to add them.
* run-test runs several ODP components such as PKTIO/Classification/crypto etc. This list located at test/linux-musdk/Makefile.am.

Change-Id: I7dfc1a7559cd77a9b6f39972d5556f7707b9dd34
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/40560
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>