mvpp2/mvgiu: fixed dependency in pktio closure sequence
fixed A7K8K-3478

The root cause was in packets that were released to HW-pool although the
pktio owner was already closed. this scenario appears when sending traffic
either between mvpp2 to mvpp2 or mvpp2 to mvgiu. The following was done:
1. for each pkt that was saved in the tx-shadow queue, its corresponding
input pktio is also saved.
2. before the saved pkt supposed to be returned to it HW-pool, its input
pktio state is checked to see if it was closed or not. if yes, the pkt is
returned to the ODP-Pool.

Change-Id: Ib594c6054ef221740c581ac6d1b6ac595efd035b
Signed-off-by: Liron Himi <lironh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/45151
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/45555